restore some case studies (#947)
* restored multiband case studies

* some updates to multi_source_per_tile case study

* added union method and removed outdated comments

* starting to unite layers in encoder

* moved train, generate, and predict into one file

* m2 case study: train with one-tile border, predict with three-tile border

* update some case studies to account for syntax changes

* log_metrics should use the encoder's sample method

* combine losses

* pylint
