+++ bliss/conf/base_config.yaml
@@ -162,7 +162,6 @@ generate:
 train:
     name: null
     version: null
-    n_epochs: 50
     save_top_k: 1
     enable_early_stopping: true
     patience: 10
@@ -175,8 +174,8 @@ train:
         reload_dataloaders_every_n_epochs: 0
         check_val_every_n_epoch: 1
         log_every_n_steps: 10  # corresponds to n_batches
-        max_epochs: ${train.n_epochs}
         min_epochs: 1
+        max_epochs: 50
         accelerator: "gpu"
         devices: 1
     testing: true
