+++ case_studies/decals_e2e_871/config_train.yaml
@@ -28,7 +28,6 @@ encoder:
 train:
     name: "pytest_encoder"
     version: "version0"
-    n_epochs: 1
     enable_early_stopping: true
     pretrained_weights: null
     seed: 42
@@ -41,4 +40,5 @@ train:
         log_every_n_steps: 1
         accelerator: "cpu"
         devices: 1
+        max_epochs: 1
     testing: true
