+++ case_studies/dependent_tiling/config.yaml
@@ -10,7 +10,7 @@ encoder:
     tile_slen: 2
     min_flux_threshold: 1.59  # 22 magnitude (0.63 is 23 mag; 0.25 is 24 mag)
     two_layers: true
-    tiles_to_crop: 3
+    tiles_to_crop: 1
 
 prior:
     _target_: bliss.simulator.prior.CatalogPrior
