+++ case_studies/multiband/config.yaml
