--- case_studies/summer_template/config.yaml
@@ -1,25 +1,0 @@
----
-defaults:
-    - ../../bliss/conf@_here_: base_config
-    - _self_
-    - override hydra/job_logging: stdout
-
-paths:
-    project: ${paths.root}/case_studies/summer_template
-    output: ${paths.project}/output
-    pretrained_models: ${paths.data}/pretrained_models
-
-simulator:
-    prior:
-        max_sources: 1
-        mean_sources: 0.02
-        min_sources: 0
-    background:
-        _target_: bliss.simulator.background.SimulatedSDSSBackground
-
-train:
-    # name: "encoder"
-    # version: "sdss_one_band"
-    # pretrained_weights: ${paths.pretrained_models}/crowded_field.pt
-    weight_save_path: ${paths.output}/${train.name}/${train.version}.pt
-    seed: 42
