--- case_studies/unalign_realign/config.yaml
@@ -1,8 +1,0 @@
----
-defaults:
-    - ../../bliss/conf@_here_: base_config
-    - _self_
-    - override hydra/job_logging: stdout
-
-simulator:
-    survey: ${surveys.sdss}
