--- data/pretrained_models/clahed_logged_2percent.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:53a3081318bf97fc98bd23768a1f1b739586e62c08981292b5113d1fc3cf4c3b
-size 55749181
